requirements.txt
# Core FastAPI
fastapi==0.110.0
uvicorn==0.27.0
python-multipart==0.0.9
python-dotenv==1.0.0

# LangChain ecosystem
langchain>=1.2.0
langchain-google-genai>=4.1.0
langchain-community>=0.4.0

# Vector store
chromadb==0.4.22

# Document loaders
pypdf==4.0.1
python-docx==1.1.0

# Utils
pydantic>=2.7.0
pydantic-settings>=2.1.0

# Google Generative AI
google-genai>=1.56.0